+++ packages/component-header-footer/setupTests.js
@@ -6,4 +6,9 @@ import "@testing-library/jest-dom";
 
 import "./__mocks__/window-mock";
 
+if (typeof global.TextEncoder === 'undefined') {
+  const util = require('util');
+  global.TextEncoder = util.TextEncoder;
+}
+
 expect.extend({ toMatchImageSnapshot });